chore(web):  splits 3d engine in its own production chunk to speed up initial page load
