+++ TODO.md
@@ -2,7 +2,6 @@
 
 ## Road to beta
 
-- improve dev loading perf (lazy-load 3d?, material icon font? https://github.com/marella/material-design-icons/tree/main/svg#readme)
 - interface in English
 - server logging (warning on invalid descriptors) + log file rotation
 - automerge.js
