+++ apps/web/src/3d/behaviors/names.js
@@ -1,4 +1,6 @@
 // we use constants instead of static class fields to break cyclic dependencies.
+// Keep this file free from @babylon (in)direct imports, to allow Svelte component referencing them
+// Otherwise this would bloat production chunks with Babylonjs (1.6Mb uncompressed)
 export const AnchorBehaviorName = 'anchorable'
 export const AnimateBehaviorName = 'animatable'
 export const DetailBehaviorName = 'detailable'
