+++ apps/web/src/routes/home/+page.svelte
@@ -53,16 +53,16 @@
     toastError({ content: translateError($_, data.creationError) })
   }
 
-  beforeNavigate(({ to }) => {
+  beforeNavigate(async ({ to }) => {
     if (user && isDifferentGame(to?.params?.gameId)) {
-      leaveGame(user)
+      await leaveGame(user)
     }
   })
 
-  function handleSelectGame({ detail: game }) {
+  async function handleSelectGame({ detail: game }) {
     if (isLobby(game)) {
       if (game.id !== $currentGame?.id) {
-        leaveGame(user)
+        await leaveGame(user)
         enterGame(game)
       }
     } else {
@@ -114,8 +114,8 @@
     })
   }
 
-  function handleCloseLobby() {
-    leaveGame(user)
+  async function handleCloseLobby() {
+    await leaveGame(user)
   }
 </script>
