+++ apps/web/src/stores/index.js
@@ -1,12 +1,13 @@
+// Do not include any store that imports anything from @src/3d (like game-engine or indicators)
+// This would bloat production chunks with Babylonjs (1.6Mb uncompressed)
+// @src/3d/utils/actions and @src/3d/behavior/names are safe
 export * from './catalog'
 export * from './configuration'
 export * from './discussion'
 export * from './friends'
 export * from './fullscreen'
-export * from './game-engine'
 export * from './game-manager'
 export * from './graphql-client'
-export * from './indicators'
 export * from './locale'
 export * from './notifications'
 export * from './peer-channels'
