+++ apps/web/src/utils/index.js
@@ -1,8 +1,10 @@
+// Do not include any store that imports anything from @src/3d (like game-interfaction)
+// This would bloat production chunks with Babylonjs (1.6Mb uncompressed)
+// @src/3d/utils/actions and @src/3d/behavior/names are safe
 export * from './collections'
 export * from './dom'
 export * from './env'
 export * from './game'
-export * from './game-interaction'
 export * from './logger'
 export * from './math'
 export * from './object'
